PDF 페이지와 내부 컴포넌트 구분
페이지는 Paper size 로 진행
컴포넌트의 크기는 PagePadding 에 따라 패딩이 적용된 위치부터 measure 진행

+++ android-pdf-writer/src/main/java/com/hangyeolee/androidpdfwriter/utils/Border.java
@@ -86,7 +86,7 @@ public class Border {
     }
 
     public void draw(BinarySerializer page, StringBuilder content, float measureX, float measureY, float measureWidth, float measureHeight){
-        int pageHeight = page.getPageHeight();
+        float pageHeight = page.getPageHeight();
         // 그래픽스 상태 저장
         content.append("q\n"); // Save graphics state
